Clang Compile Fix

+++ src/hir/deserialise.cpp
@@ -721,7 +721,7 @@ namespace {
 
     ::HIR::LifetimeRef HirDeserialiser::deserialise_lifetimeref()
     {
-        return { m_in.read_count() };
+        return { static_cast<uint32_t>(m_in.read_count()) };
     }
 
     ::HIR::TypeRef HirDeserialiser::deserialise_type()
